13.0 dev

+++ dingtalk_hr/models/hr_employee.py
@@ -149,7 +149,7 @@ class HrEmployee(models.Model):
 
     def unlink(self):
         for res in self:
-            if res.ding_id:
+            if res.ding_id and dingtalk_api.get_delete_is_synchronous():
                 self._delete_dingtalk_employee_by_id(res.ding_id)
             return super(HrEmployee, self).unlink()
